Syntactic sugar and wrapper code generation framework.
Currently Matcher factory methods are defined all over the place which makes it really hard to remember which methods need to be imported and doesn't allow you to easily browse a list of all of them. This generates one uber matcher factory that delegates to all the respective classes. 

Additionally, this allows the uber-classes to do things like wrap matchers in adaptors (e.g. to make them EasyMock friendly).

This also allows users to group their own collections of Matchers together in a way that makes sense to them.